add file package
