+++ pyzoo/zoo/orca/data/utils.py
@@ -87,77 +87,3 @@ def extract_one_path(file_path, file_type, env):
             file_paths = [file for file in file_paths
                           if os.path.splitext(file)[1] == "." + file_type]
     return file_paths
-
-
-def open_text(path):
-    # Return a list of lines
-    if path.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        fs = pa.hdfs.connect()
-        with fs.open(path, 'rb') as f:
-            lines = f.read().decode("utf-8").split("\n")
-    elif path.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = path.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        key = "/".join(path_parts)
-        data = s3_client.get_object(Bucket=bucket, Key=key)
-        lines = data["Body"].read().decode("utf-8").split("\n")
-    else:  # Local path
-        lines = []
-        for line in open(path):
-            lines.append(line)
-    return [line.strip() for line in lines]
-
-
-def open_image(path):
-    from PIL import Image
-    if path.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        fs = pa.hdfs.connect()
-        with fs.open(path, 'rb') as f:
-            return Image.open(f)
-    elif path.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        from io import BytesIO
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = path.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        key = "/".join(path_parts)
-        data = s3_client.get_object(Bucket=bucket, Key=key)
-        return Image.open(BytesIO(data["Body"].read()))
-    else:  # Local path
-        return Image.open(path)
-
-
-def load_numpy(path):
-    import numpy as np
-    if path.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        fs = pa.hdfs.connect()
-        with fs.open(path, 'rb') as f:
-            return np.load(f)
-    elif path.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        from io import BytesIO
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = path.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        key = "/".join(path_parts)
-        data = s3_client.get_object(Bucket=bucket, Key=key)
-        return np.load(BytesIO(data["Body"].read()))
-    else:  # Local path
-        return np.load(path)
